requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.1.3
fasteners==0.14.1
frozenlist==1.4.0
future==0.18.2
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-speech==2.22.0
google-cloud-texttospeech==2.14.2
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.25.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openai==0.28.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
playsound==1.3.0
proto-plus==1.22.3
protobuf==4.25.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycairo==1.20.1
pycups==2.0.1
pydantic==2.5.0
pydantic_core==2.14.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
systemd-python==234
tqdm==4.66.1
typing_extensions==4.8.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
Adafruit-Blinka==8.25.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.10
adafruit-circuitpython-pixelbuf==2.0.3
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.54.0
Adafruit-PureIO==1.1.11
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.1.3
fasteners==0.14.1
frozenlist==1.4.0
future==0.18.2
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-speech==2.22.0
google-cloud-texttospeech==2.14.2
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.25.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
openai==0.28.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
playsound==1.3.0
proto-plus==1.22.3
protobuf==4.25.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.5.0
pydantic_core==2.14.1
pyftdi==0.55.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyusb==1.2.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
scipy==1.11.4
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
systemd-python==234
tqdm==4.66.1
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0